# web/openapi.json
{
  "components": {
    "schemas": {
      "ActionItem": {
        "properties": {
          "assignee_agent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Agent Id"
          },
          "assignee_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee Name"
          },
          "decision_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "task_id": {
            "title": "Task Id",
            "type": "integer"
          },
          "task_status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "task_title": {
            "title": "Task Title",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text",
          "decision_id",
          "assignee_agent_id",
          "assignee_name",
          "task_id",
          "task_title",
          "task_status"
        ],
        "title": "ActionItem",
        "type": "object"
      },
...
      "ApprovalStatus": {
        "enum": ["pending", "approved", "rejected", "executed", "execution_failed", "cancelled"],
        "title": "ApprovalStatus",
        "type": "string"
      },
...
      "Decided": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/DecidedAction"
            },
            "title": "Actions",
            "type": "array"
          },
          "decisions": {
            "items": {
              "type": "string"
            },
            "title": "Decisions",
            "type": "array"
          },
          "ended_at": {
            "format": "date-time",
            "title": "Ended At",
            "type": "string"
          },
          "meeting_id": {
            "title": "Meeting Id",
            "type": "integer"
          },
          "pinned_id": {
            "title": "Pinned Id",
            "type": "integer"
          },
          "pinned_kind": {
            "title": "Pinned Kind",
            "type": "string"
          }
        },
        "required": ["meeting_id", "pinned_kind", "pinned_id", "decisions", "actions", "ended_at"],
        "title": "Decided",
        "type": "object"
      },
      "DecidedAction": {
        "properties": {
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "approval_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["text", "approval_id", "approval_status"],
        "title": "DecidedAction",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": ["error"],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "MeetingDetail": {
        "properties": {
          "action_items": {
            "items": {
              "$ref": "#/components/schemas/ActionItem"
            },
            "title": "Action Items",
            "type": "array"
          },
          "agenda": {
            "title": "Agenda",
            "type": "string"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/labhq__api__meetings__schemas__Decision"
            },
            "title": "Decisions",
            "type": "array"
          },
          "end_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Reason"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Messages",
            "type": "array"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/Participant"
            },
            "title": "Participants",
            "type": "array"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          }
        },
        "required": [
          "id",
          "kind",
          "project_id",
          "project_name",
          "status",
          "created_at",
          "started_at",
          "channel",
          "agenda",
          "end_reason",
          "ended_at",
          "participants",
          "messages",
          "decisions",
          "action_items"
        ],
        "title": "MeetingDetail",
        "type": "object"
      },
...
      "MeetingStatus": {
        "enum": ["requested", "running", "ended", "failed", "cancelled"],
        "title": "MeetingStatus",
        "type": "string"
      },
      "Message": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/TranscriptSource"
          },
          "speaker": {
            "title": "Speaker",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["id", "source", "speaker", "agent_id", "text", "created_at"],
        "title": "Message",
        "type": "object"
      },
...
      "Participant": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": ["id", "agent_id", "name"],
        "title": "Participant",
        "type": "object"
      },
...
      "RoomItem": {
        "properties": {
          "agenda": {
            "title": "Agenda",
            "type": "string"
          },
          "approval_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Id"
          },
          "approval_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "end_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Reason"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "estimate_micros": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimate Micros"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "pinned_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Id"
          },
          "pinned_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned Kind"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          },
          "project_name": {
            "title": "Project Name",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MeetingStatus"
          },
          "turn_cap": {
            "title": "Turn Cap",
            "type": "integer"
          },
          "turns_used": {
            "title": "Turns Used",
            "type": "integer"
          },
          "waiting": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Waiting"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "status",
          "project_id",
          "project_name",
          "agenda",
          "pinned_kind",
          "pinned_id",
          "estimate_micros",
          "approval_id",
          "approval_status",
          "turns_used",
          "turn_cap",
          "waiting",
          "end_reason",
          "created_at",
          "ended_at"
        ],
        "title": "RoomItem",
        "type": "object"
      },
      "RoomMessage": {
        "properties": {
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["text"],
        "title": "RoomMessage",
        "type": "object"
      },
...
      "TaskStatus": {
        "enum": ["backlog", "todo", "in_progress", "in_review", "blocked", "done", "cancelled"],
        "title": "TaskStatus",
        "type": "string"
      },
...
      "TranscriptSource": {
        "enum": ["agent", "owner", "system"],
        "title": "TranscriptSource",
        "type": "string"
      },
...
      "Waiting": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "integer"
          },
          "agent_name": {
            "title": "Agent Name",
            "type": "string"
          },
          "can_interrupt": {
            "title": "Can Interrupt",
            "type": "boolean"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": ["agent_id", "agent_name", "reason", "can_interrupt"],
        "title": "Waiting",
        "type": "object"
      },
...
      "labhq__api__meetings__schemas__Decision": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": ["id", "position", "text"],
        "title": "Decision",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "labhq",
    "version": "0.1.0a1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/callcenter/decisions": {
      "get": {
        "description": "What each ended room decided about the proposal it was pinned to, for the cards.",
        "operationId": "decisions_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Decided"
                  },
                  "title": "Response Decisions List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Decisions List",
        "tags": ["callcenter"]
      }
    },
    "/api/callcenter/rooms": {
      "get": {
        "description": "The newest rooms, newest first; the widget opens the first that is not over.",
        "operationId": "rooms_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoomItem"
                  },
                  "title": "Response Rooms List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Rooms List",
        "tags": ["callcenter"]
      }
    },
    "/api/callcenter/rooms/{room_id}/close": {
      "post": {
        "description": "Write the minutes and end the room, in the background.",
        "operationId": "room_close",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Room Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Room Close",
        "tags": ["callcenter"]
      }
    },
    "/api/callcenter/rooms/{room_id}/decline": {
      "post": {
        "description": "Refuse the start; the room is cancelled and nobody is asked.",
        "operationId": "room_decline",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Room Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Room Decline",
        "tags": ["callcenter"]
      }
    },
    "/api/callcenter/rooms/{room_id}/messages": {
      "post": {
        "description": "Say something as the owner. It is recorded at once; the agents answer in the background.",
        "operationId": "room_say",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Room Id",
              "type": "integer"
            }
          },
          {
            "description": "One per user intent; a retry reuses it.",
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "One per user intent; a retry reuses it.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomMessage"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Room Say",
        "tags": ["callcenter"]
      }
    },
    "/api/callcenter/rooms/{room_id}/start": {
      "post": {
        "description": "The owner's tap on the room's start approval, then the room opens in the background.",
        "operationId": "room_start",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Room Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Room Start",
        "tags": ["callcenter"]
      }
    },
...
    "/api/meetings/{meeting_id}": {
      "get": {
        "description": "The whole meeting from the database alone, so it reads the same with chat disconnected.",
        "operationId": "meetings_get",
        "parameters": [
          {
            "in": "path",
            "name": "meeting_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Meeting Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Meetings Get",
        "tags": ["meetings"]
      }
    },
...
  }
}